requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
pinecone-client==3.0.2
python-dotenv==1.0.1
huggingface-hub==0.21.4
chromadb==0.4.24
